requirements.txt
streamlit
tensorflow
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
Pillow
